+++ src/components/Select/Select.tsx
@@ -1,103 +1,58 @@
-import Select, {
-  components,
-  DropdownIndicatorProps,
-  GroupBase,
-  IndicatorSeparatorProps,
-  StylesConfig,
-} from 'react-select';
+import { useRef, useState } from 'react';
+import { useClickAway } from 'react-use';
+import { ArrowDownSVG } from '../../../public/icons/svg';
 
-const { DropdownIndicator, IndicatorSeparator } = components;
-
-const CustomDropdownIndicator = (
-  props: DropdownIndicatorProps<{ value: number; label: string }, false>,
-) => {
-  return props.selectProps.menuIsOpen ? null : <DropdownIndicator {...props} />;
-};
-
-const CustomIndicatorSeparator = (
-  props: IndicatorSeparatorProps<{ value: number; label: string }, false>,
-) => {
-  return props.selectProps.menuIsOpen ? null : (
-    <IndicatorSeparator {...props} />
-  );
-};
-
-interface SelectComponentProps {
-  instanceId?: string;
-  defaultValue?: { value: number; label: string };
-  onChange?: (selectedOption: { value: number; label: string } | null) => void;
-  isDisabled?: boolean;
-  options?: { value: number; label: string }[];
+interface SelectProps {
+  defaultValue: string;
+  selectList: string[];
+  onChange?: () => void;
 }
 
-const SelectComponent = ({
-  instanceId,
-  defaultValue,
-  onChange,
-  isDisabled = false,
-  options = [],
-}: SelectComponentProps) => {
+const Select = ({ defaultValue, selectList, onChange }: SelectProps) => {
+  const [selectedValue, setSelectedValue] = useState(defaultValue);
+  const [isDropdownVisible, setDropdownVisibility] = useState(false);
+  const selectRef = useRef(null);
+
+  useClickAway(selectRef, () => {
+    setDropdownVisibility(false);
+  });
+
+  const handleSelectionChange = (value: string) => {
+    setSelectedValue(value);
+  };
+
+  const toggleDropdownVisibility = () => {
+    setDropdownVisibility((prev) => !prev);
+  };
+
   return (
-    <Select
-      instanceId={instanceId}
-      defaultValue={defaultValue}
-      onChange={onChange}
-      isDisabled={isDisabled}
-      options={options}
-      components={{
-        DropdownIndicator: CustomDropdownIndicator,
-        IndicatorSeparator: CustomIndicatorSeparator,
-      }}
-      styles={selectStyles}
-    />
+    <div
+      ref={selectRef}
+      onClick={toggleDropdownVisibility}
+      className={`relative flex cursor-pointer items-center ${
+        isDropdownVisible &&
+        'rounded-t-md shadow-[1px_-1px_4px_0px_rgba(0,0,0,0.25)]'
+      }`}
+    >
+      {selectedValue}
+      <ArrowDownSVG className="h-6 w-6 -translate-x-1.5 -translate-y-1.5 fill-sub-color2" />
+      {isDropdownVisible && (
+        <ul className="absolute left-0 right-0 top-full rounded-b-md bg-white shadow-[1px_1px_4px_0px_rgba(0,0,0,0.25)]">
+          {selectList.map((value) =>
+            selectedValue === value ? null : (
+              <li
+                key={value}
+                className="border-t border-solid border-sub-color2"
+                onClick={() => handleSelectionChange(value)}
+              >
+                {value}
+              </li>
+            ),
+          )}
+        </ul>
+      )}
+    </div>
   );
 };
 
-export default SelectComponent;
-
-const selectStyles:
-  | StylesConfig<
-      {
-        value: number;
-        label: string;
-      },
-      false,
-      GroupBase<{
-        value: number;
-        label: string;
-      }>
-    >
-  | undefined = {
-  menuList: (base) => ({ ...base, maxHeight: '200px' }),
-  control: (provided, state) => ({
-    ...provided,
-    borderRadius:
-      state.selectProps.menuIsOpen || state.menuIsOpen
-        ? '1.3rem 1.3rem 0 0'
-        : '1.875rem',
-    boxShadow: 'none',
-    borderColor: 'var(--sub-color4)',
-    '&:hover': { borderColor: 'var(--sub-color4)' },
-  }),
-  valueContainer: (provided, state) => ({
-    ...provided,
-    width: state.selectProps.menuIsOpen ? '4.275rem' : provided.width,
-    justifyContent: 'center',
-    borderColor: 'none',
-  }),
-  menu: (provided) => ({
-    ...provided,
-    marginTop: '-10px',
-    boxShadow: 'none',
-    border: '1px solid var(--sub-color4)',
-    borderTop: 'none',
-  }),
-  option: (provided, state) => ({
-    ...provided,
-    backgroundColor: state.isSelected ? 'var(--sub-color4)' : 'white',
-    color: 'black',
-    '&:hover': {
-      backgroundColor: '#eceaea',
-    },
-  }),
-};
+export default Select;
